refactor(validator): small improvements
