+++ feature/validator/src/commonMain/kotlin/com/neoutils/neorefex/feature/validator/ValidatorScreen.kt
@@ -55,6 +55,7 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import cafe.adriel.voyager.core.screen.Screen
 import cafe.adriel.voyager.koin.koinScreenModel
 import com.neoutils.neorefex.feature.validator.action.ValidatorAction
+import com.neoutils.neorefex.feature.validator.state.TestCase
 import com.neoutils.neoregex.core.common.model.TestCase
 import com.neoutils.neorefex.feature.validator.state.ValidatorUiState
 import com.neoutils.neoregex.core.designsystem.component.ErrorTooltip
@@ -213,285 +214,5 @@ class ValidatorScreen : Screen {
             },
             onAction = viewModel::onAction
         )
-    }
-}
-
-@OptIn(
-    ExperimentalMaterial3Api::class,
-    ExperimentalUuidApi::class
-)
-@Composable
-private fun TestCase(
-    test: TestCase,
-    expanded: Boolean,
-    onTestChange: (TestCase) -> Unit,
-    modifier: Modifier = Modifier,
-    onExpanded: () -> Unit = {},
-    onClose: () -> Unit = {},
-    onDelete: () -> Unit = {},
-    onCopy: () -> Unit = {},
-    textStyle: TextStyle = TextStyle(),
-    contentPadding: PaddingValues = PaddingValues(dimensions.default),
-    hint: String = "Enter input"
-) {
-    val color by animateColorAsState(
-        when (test.result) {
-            TestCase.Result.IDLE -> colorScheme.outlineVariant
-
-            TestCase.Result.RUNNING -> {
-                rememberInfiniteTransition().animateColor(
-                    initialValue = colorScheme.outlineVariant,
-                    targetValue = Color.White,
-                    animationSpec = infiniteRepeatable(
-                        animation = tween(durationMillis = 1000),
-                        repeatMode = RepeatMode.Reverse
-                    )
-                ).value
-            }
-
-            TestCase.Result.SUCCESS -> Green
-            TestCase.Result.ERROR -> colorScheme.error
-        }
-    )
-
-    Surface(
-        modifier = modifier,
-        shape = RoundedCornerShape(4.dp),
-        color = colorScheme.surfaceContainer,
-        contentColor = colorScheme.onSurface,
-        border = BorderStroke(
-            width = 1.dp,
-            color = color
-        ),
-    ) {
-        Column {
-            AnimatedVisibility(visible = expanded) {
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.spacedBy(dimensions.small),
-                    modifier = Modifier
-                        .padding(horizontal = 8.dp)
-                        .padding(top = 8.dp)
-                        .fillMaxWidth(),
-                ) {
-                    NeoTextField(
-                        value = test.title,
-                        onValueChange = {
-                            onTestChange(
-                                test.copy(
-                                    title = it
-                                )
-                            )
-                        },
-                        contentPadding = PaddingValues(0.dp),
-                        textStyle = typography.labelMedium,
-                        hint = "Untitled",
-                        singleLine = true,
-                        modifier = Modifier
-                            .padding(start = 8.dp)
-                            .fillMaxWidth()
-                            .weight(1f)
-                    )
-
-                    Options(
-                        case = test.case,
-                        onCaseChange = {
-                            onTestChange(
-                                test.copy(
-                                    case = it
-                                )
-                            )
-                        },
-                        onDelete = onDelete,
-                        onClose = onClose,
-                        onCopy = onCopy
-                    )
-                }
-            }
-
-            val mergedTextStyle = typography.bodyLarge.merge(textStyle)
-
-            AnimatedContent(
-                targetState = expanded,
-                transitionSpec = {
-                    fadeIn() togetherWith fadeOut()
-                }
-            ) { expanded ->
-                if (expanded) {
-
-                    val focusRequester = remember { FocusRequester() }
-
-                    LaunchedEffect(Unit) {
-                        focusRequester.requestFocus()
-                    }
-
-                    NeoTextField(
-                        value = test.text,
-                        onValueChange = {
-                            onTestChange(
-                                test.copy(
-                                    text = it
-                                )
-                            )
-                        },
-                        modifier = Modifier
-                            .focusRequester(focusRequester)
-                            .fillMaxWidth(),
-                        contentPadding = contentPadding,
-                        textStyle = mergedTextStyle,
-                        hint = hint
-                    )
-                } else {
-
-                    val title = test.title.ifEmpty {
-                        test.text.ifEmpty {
-                            "Untitled"
-                        }
-                    }
-
-                    TextFieldDefaults.DecorationBox(
-                        value = title,
-                        innerTextField = {
-                            Text(
-                                text = title,
-                                modifier = Modifier.fillMaxWidth(),
-                                style = mergedTextStyle,
-                                overflow = TextOverflow.Ellipsis,
-                                maxLines = 1
-                            )
-                        },
-                        enabled = true,
-                        singleLine = true,
-                        visualTransformation = VisualTransformation.None,
-                        interactionSource = remember { MutableInteractionSource() },
-                        contentPadding = contentPadding,
-                        container = {
-                            Box(
-                                modifier.clickable(
-                                    onClick = onExpanded
-                                )
-                            )
-                        },
-                        trailingIcon = {
-                            Text(
-                                text = test.case.text,
-                                modifier = Modifier.padding(contentPadding),
-                                style = typography.labelMedium
-                            )
-                        }
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun Options(
-    case: TestCase.Case,
-    onCaseChange: (TestCase.Case) -> Unit,
-    modifier: Modifier = Modifier,
-    onDelete: () -> Unit = {},
-    onClose: () -> Unit = {},
-    onCopy: () -> Unit = {},
-) {
-    Row(
-        modifier = modifier,
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(dimensions.tiny)
-    ) {
-        MatchDropDown(
-            case = case,
-            onChange = {
-                onCaseChange(it)
-            }
-        )
-
-        IconButton(
-            onClick = onCopy,
-            modifier = Modifier.size(24.dp)
-        ) {
-            Icon(
-                imageVector = Icons.Rounded.ContentCopy,
-                contentDescription = null,
-                modifier = Modifier
-                    .padding(vertical = 0.5.dp)
-                    .padding(4.dp)
-            )
-        }
-
-        IconButton(
-            onClick = onDelete,
-            modifier = Modifier.size(24.dp)
-        ) {
-            Icon(
-                imageVector = Icons.Outlined.Delete,
-                contentDescription = null,
-                modifier = Modifier.padding(4.dp)
-            )
-        }
-
-        IconButton(
-            onClick = onClose,
-            modifier = Modifier.size(24.dp)
-        ) {
-            Icon(
-                imageVector = Icons.Outlined.ExpandCircleDown,
-                contentDescription = null,
-                modifier = Modifier.padding(4.dp)
-            )
-        }
-    }
-}
-
-@Composable
-private fun MatchDropDown(
-    case: TestCase.Case,
-    onChange: (TestCase.Case) -> Unit
-) {
-    Column(verticalArrangement = Arrangement.Center) {
-
-        val expanded = remember { mutableStateOf(false) }
-
-        Link(
-            text = case.text,
-            onClick = {
-                expanded.value = true
-            },
-            enabledUnderline = false,
-            colors = LinkColor(
-                idle = colorScheme.onSurface,
-                pressed = colorScheme.onSurface,
-                press = colorScheme.onSurface.copy(alpha = 0.6f),
-                hover = colorScheme.onSurface.copy(alpha = 0.8f)
-            ),
-            endIcon = {
-                Icon(
-                    imageVector = Icons.Outlined.KeyboardArrowDown,
-                    modifier = Modifier.size(18.dp),
-                    contentDescription = null
-                )
-            },
-            style = typography.labelMedium
-        )
-
-        DropdownMenu(
-            expanded = expanded.value,
-            onDismissRequest = { expanded.value = false }
-        ) {
-            TestCase.Case.entries.forEach {
-                DropdownMenuItem(
-                    text = {
-                        Text(
-                            text = it.text,
-                        )
-                    },
-                    onClick = {
-                        expanded.value = false
-                        onChange(it)
-                    },
-                )
-            }
-        }
     }
 }
